backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
pydantic[email]>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
qrcode[pil]>=7.4.2
openpyxl>=3.1.2
reportlab>=4.2.0

# Testing
pytest>=8.2.0
pytest-cov>=5.0.0
httpx>=0.27.0
